refactor(tui): move plugin render components to render.tsx
